refactor(vtc): move join-submit orchestration into crate::join (P2.1, part 3)
The join `decide → auto-admit-effect → audit` spine — `submit_inner`,
`decide_join`, `realize_join_verdict`, `assemble_join_facts`, the holder-binding
verification + canonical-payload helpers, the freshness consts, and the
`HolderBinding` / `JoinSubmitOutcome` types — lived in
`routes/join_requests/submit.rs` but is shared by three entry points (REST
submit, the DIDComm handler in `messaging.rs`, and credential-exchange
`present.rs`). Move it into a new `join::orchestrate` module so it lives beside
the join lifecycle it persists (`JoinRequest`, `store_join_request`) and no
caller reaches back through `crate::routes::...`.

The shared admit-audit helper (`emit_admit_audit` + `credential_issued_data`,
the `MemberAdded`/`VmcIssued`/`VecIssued` envelopes) moves here too — it was in
`routes/join_requests/audit.rs`, used by both the auto-admit path and the
manual-approve route (`decide::approve`); that route now calls
`crate::join::emit_admit_audit`.

`routes/join_requests/submit.rs` is now a thin adapter: the `SubmitRequestBody`
/ `SubmitResponse` wire types + the `submit` handler (auth/body extraction →
`crate::join::submit_inner` → response shaping). 824 → 153 lines.

This completes P2.1 — all three ceremonies (role-change #459, leave #460, join)
plus the admit-audit emission now live outside the route layer; handlers carry
no policy-load / decide / effect / audit logic. Behaviour-preserving: the moved
holder-signature round-trip unit tests pass in their new home; join_requests
(46), join_didcomm (1), full lib (582) green; fmt + clippy clean.
