+++ vtc-service/src/messaging.rs
@@ -31,11 +31,11 @@ use vta_sdk::protocols::join_requests::{
 use crate::ceremony::remove_inner;
 use crate::config::AppConfig;
 use crate::join::JoinTransport;
+use crate::join::submit_inner;
 use crate::members::Disposition;
 use crate::routes::join_requests::accept::accept_inner;
 use crate::routes::join_requests::manifest::manifest_inner;
 use crate::routes::join_requests::status::status_inner;
-use crate::routes::join_requests::submit::submit_inner;
 use crate::server::AppState;
 
 /// Start the DIDComm service and block until shutdown.
